Addressing comments from software WG: names, commented code blocks. All pfRICH gdml files are removed
