--- compact/pid/pfrich_gdml.xml
@@ -1,8 +1,0 @@
-<lccdd>
-  <detectors>
-    <detector id="BackwardRICH_ID" name="PFRICH" type="DD4hep_GdmlDetector">
-      <gdmlFile ref="pfrich.gdml"/>
-      <parent name="/world_volume"/>
-    </detector>
-  </detectors>
-</lccdd>
