--- epic_pfrich_sl_only.xml
@@ -1,121 +1,0 @@
-<lccdd xmlns:compact="http://www.lcsim.org/schemas/compact/1.0"
-       xmlns:xs="http://www.w3.org/2001/XMLSchema"
-       xs:noNamespaceSchemaLocation="http://www.lcsim.org/schemas/compact/1.0/compact.xsd">
-
-  <debug>
-    <type name="surface"       value="0"/>
-    <type name="material"      value="0"/>
-    <type name="readout"       value="0"/>
-    <type name="segmentation"  value="0"/>
-    <type name="limits"        value="0"/>
-    <type name="region"        value="0"/>
-    <type name="includes"      value="0"/>
-  </debug>
-
-  <documentation level="-1">
-  # EPIC Detector
-  - https://github.com/eic/epic
-  - https://github.com/eic/ip6
-  </documentation>
-
-  <!-- Some information about detector  -->
-  <info name="EPIC Detector" title="EPIC Detector"
-        author="EPIC Collaboration"
-        url="https://github.com/eic/epic"
-        status="development"
-        version="v1 2021-03-16">
-  <comment> EPIC </comment>
-  </info>
-
-  <!-- Standard environmental conditions: 20 °C, 1 atm -->
-  <std_conditions type="NTP"/>
-
-  <define>
-    <documentation level="2">
-      ## Main Constant Definitions
-
-      The ip6 (or other ip) defines should be included first.
-      These files have only a define tags.
-    </documentation>
-    <include ref="${DETECTOR_PATH}/compact/fields/beamline_18x275.xml" />
-    <include ref="${DETECTOR_PATH}/compact/far_forward/definitions.xml" />
-    <include ref="${DETECTOR_PATH}/compact/far_backward/definitions.xml" />
-    <include ref="${DETECTOR_PATH}/compact/definitions.xml" />
-    <include ref="${DD4hepINSTALL}/DDDetectors/compact/detector_types.xml"/>
-  </define>
-
-
-  <includes>
-    <gdmlFile ref="${DD4hepINSTALL}/DDDetectors/compact/elements.xml"/>
-    <gdmlFile ref="${DETECTOR_PATH}/compact/materials.xml"/>
-  </includes>
-  <include ref="${DETECTOR_PATH}/compact/optical_materials.xml"/>
-
-  <limits>
-    <limitset name="world_limits">
-    </limitset>
-    <limitset name="beamline_limits">
-      <limit name="step_length_max" particles="*" value="1.0" unit="mm" />
-      <limit name="track_length_max" particles="*" value="1.0" unit="mm" />
-      <limit name="time_max" particles="*" value="0.1" unit="ns" />
-      <limit name="ekin_min" particles="*" value="0.001" unit="MeV" />
-      <limit name="range_min" particles="*" value="0.1" unit="mm" />
-    </limitset>
-    <limitset name="cal_limits">
-      <limit name="step_length_max" particles="*" value="5.0" unit="mm"/>
-    </limitset>
-  </limits>
-
-  <regions>
-    <region name="world_region" eunit="MeV" lunit="mm" cut="0.001" threshold="0.001">
-      <limitsetref name="world_limits"/>
-    </region>
-  </regions>
-
-  <display>
-    <include ref="${DETECTOR_PATH}/compact/colors.xml"/>
-    <include ref="${DETECTOR_PATH}/compact/display.xml"/>
-  </display>
-
-  <documentation level="0">
-    ## World Volume
-
-    The world is a simple box, but could be a union of multiple regions.
-  </documentation>
-  <world material="Air">
-    <shape type="Box" dx="world_dx" dy="world_dy" dz="world_dz"/>
-    <!--regionref   name="world_region"/-->
-    <!--limitsetref name="world_limits"/-->
-  </world>
-
-  <documentation level="0">
-    ## Detector Subsystems
-
-    ### IP Subsystems
-
-    The interaction point subsystems are included before the central detector subsystems.
-    This is becuase the IP subsystems, for example the beampipe, will define paramters
-    which are subsquently used in the central detector construction -- e.g. the vertex tracker
-    uses the beampipe OD to help define its placement.
-
-    The IP subsystems include the Far forward and backward regions. The list of subsystem includes:
-     - Interaction region beampipe
-     - B0 tracker
-     - Off-momentum tracker
-     - Far forward roman pots
-     - Zero Degree Calorimeter
-     - Beam line magnets.
-     - and more...
-  </documentation>
-
-
-  <documentation level="5">
-    ## Electric and magnetic fields
-  </documentation>
-  <include ref="${DETECTOR_PATH}/compact/fields/marco.xml"/>
-  
-  <documentation level="10">
-    ## PID detectors
-  </documentation>
-  <include ref="${DETECTOR_PATH}/compact/pid/pfrich_sl.xml"/>
-  </lccdd>
